Split up artwork.c
