+++ plugins/artwork/Makefile.am
@@ -9,9 +9,10 @@ endif
 
 if ARTWORK_USE_VFS_CURL
 artwork_net_cflags = -DUSE_VFS_CURL
+artwork_net_sources =  albumartorg.c albumartorg.h lastfm.c lastfm.h wos.c wos.h escape.c escape.h
 endif
 
-artwork_la_SOURCES = artwork.c artwork.h escape.c escape.h
+artwork_la_SOURCES = artwork.c artwork.h cache.c cache.h artwork_internal.c artwork_internal.h $(artwork_net_sources)
 
 artwork_la_LDFLAGS = -module -avoid-version
