+++ plugins/artwork/wos.h
@@ -1,0 +1,29 @@
+/*
+    world of spectrum game cover downloader for the Deadbeef Artwork plugin
+    Copyright (C) 2009-2013 Alexey Yakovenko and other contributors
+
+    This software is provided 'as-is', without any express or implied
+    warranty.  In no event will the authors be held liable for any damages
+    arising from the use of this software.
+
+    Permission is granted to anyone to use this software for any purpose,
+    including commercial applications, and to alter it and redistribute it
+    freely, subject to the following restrictions:
+
+    1. The origin of this software must not be misrepresented; you must not
+     claim that you wrote the original software. If you use this software
+     in a product, an acknowledgment in the product documentation would be
+     appreciated but is not required.
+
+    2. Altered source versions must be plainly marked as such, and must not be
+     misrepresented as being the original software.
+
+    3. This notice may not be removed or altered from any source distribution.
+*/
+#ifndef __WOS_H
+#define __WOS_H
+
+int
+fetch_from_wos (const char *title, const char *dest);
+
+#endif // __WOS_H
